tests: Move 'mocked pagers' utils to separate file
